+++ python/GafferArnoldUI/__init__.py
@@ -56,6 +56,7 @@ from . import ArnoldLightFilterUI
 from . import ArnoldColorManagerUI
 from . import ArnoldImagerUI
 from . import ArnoldProceduralUI
+from . import ArnoldOperatorUI
 from . import CacheMenu
 
 __import__( "IECore" ).loadConfig( "GAFFER_STARTUP_PATHS", subdirectory = "GafferArnoldUI" )