+++ src/IECoreArnold/Renderer.cpp
@@ -71,6 +71,7 @@
 
 #include "ai_array.h"
 #include "ai_msg.h"
+#include "ai_operator.h"
 #include "ai_procedural.h"
 #include "ai_ray.h"
 #include "ai_render.h"
@@ -1245,11 +1246,13 @@ const std::vector<IECore::InternedString> g_surfaceShaderAttributeNames = {
 	"osl:surface",
 	/// \todo Remove support for interpreting "osl:shader" as a surface shader assignment.
 	"osl:shader",
+	"mtlx:surface",
 	"surface"
 };
 
 const std::vector<IECore::InternedString> g_volumeShaderAttributeNames = {
 	"ai:volume",
+	"mtlx:volume",
 	"volume"
 };
 
@@ -3459,6 +3462,7 @@ const IECore::InternedString g_imagerOptionName( "ai:imager" );
 const IECore::InternedString g_instanceIDAOVShaderOptionName( "ai:aov_shader:__cortexInstanceID" );
 const IECore::InternedString g_logFileNameOptionName( "ai:log:filename" );
 const IECore::InternedString g_logMaxWarningsOptionName( "ai:log:max_warnings" );
+const IECore::InternedString g_operatorOptionName( "ai:operator" );
 const IECore::InternedString g_pluginSearchPathOptionName( "ai:plugin_searchpath" );
 const IECore::InternedString g_profileFileNameOptionName( "ai:profileFileName" );
 const IECore::InternedString g_progressiveMinAASamplesOptionName( "ai:progressive_min_AA_samples" );
@@ -3581,6 +3585,7 @@ class ArnoldGlobals
 			m_atmosphere.reset();
 			m_background.reset();
 			m_imager.reset();
+			m_operator.reset();
 			m_defaultCamera.reset();
 			// Destroy the universe while our message callback is
 			// still active, so we catch any Arnold shutdown messages.
@@ -3895,6 +3900,19 @@ class ArnoldGlobals
 				// NOTE : If we weren't always updating the drivers, we would need to flag here that the driver
 				// has been dirtied. However, currently, we've been unable to observe any perceptible cost to just
 				// updating the drivers every time we render, so we don't need to flag anything here.
+				return;
+			}
+			else if( name == g_operatorOptionName )
+			{
+				m_operator = nullptr;
+				if( value )
+				{
+					if( const IECoreScene::ShaderNetwork *d = reportedCast<const IECoreScene::ShaderNetwork>( value, "option", name ) )
+					{
+						m_operator = m_shaderCache->get( d, IECore::InternedString(), nullptr );
+					}
+				}
+				AiOpSetTarget( universe(), m_operator ? m_operator->root() : nullptr );
 				return;
 			}
 			else if( boost::starts_with( name.c_str(), "ai:aov_shader:" ) )
@@ -4642,6 +4660,7 @@ class ArnoldGlobals
 		ArnoldShaderPtr m_atmosphere;
 		ArnoldShaderPtr m_background;
 		ArnoldShaderPtr m_imager;
+		ArnoldShaderPtr m_operator;
 
 		std::string m_cameraName;
 		using CameraMap = tbb::concurrent_unordered_map<std::string, IECoreScene::ConstCameraPtr>;