+++ Changes.md
@@ -1113,6 +1113,11 @@ API
 1.5.4.1 (relative to 1.5.4.0)
 =======
 
+Improvements
+------------
+
+- USDLight : Append the (Renderman) suffix to any USD Lux plugs with `ri:light:` prefix.
+
 Fixes
 -----
 
@@ -1642,12 +1647,17 @@ Improvements
 ------------
 
 - Arnold : Added support for Int64Data and UInt64Data custom attributes, allowing USD's `instanceId` to be used as a custom attribute in the Instancer node. Warnings are emitted if values are out of range for Arnold's 32 bit ints.
+- USDShader :
+  - A namespace prefix is now set for the shader type from the USD source type, except for `USD` and `glslfx` built-ins which correspond to USDLux lights and USDPreviewSurface built-ins
+  - Added a way to register namespace overrides from USD shader source types if the resulting USD source type is undesirable
+  - The correct shader assignment for displacement and volume shaders are now set for shaders which have a context of `displacement` or `volume` set
 
 Fixes
 -----
 
 - SceneReader : Fixed crash reading facevarying normals skinned with UsdSkel.
 - ShaderView : Fixed crash caused by a SceneCreator returning `None`.
+- USDShader : Fixed default preset values on IntPlugs that are intended to be enum indexes instead of string values.
 
 Build
 -----